fix: Prevent treasure box persistence after completing a stage without collecting it
Added a condition (Stage1Boss_dead = 0) within the 'if(hard1start == 0)' block in the GoldBox1Thread2 thread to address the issue where the treasure box remains unaffected after completing a stage without collecting it and restarting the stage within the game. This modification ensures that the treasure box is properly reset in such scenarios.